Solve forking issue
Added the section “Estimation when subjects are assigned with varying
treatment probabilities”. This section was added to Don and Winston’s
Google Doc in June, and shared with a few colleagues in a June 24 PDF,
but hasn’t previously appeared in the GitHub version.

Thus, the current version takes care of a small forking issue and
allows us to divorce Google Docs and marry GitHub.